+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 0.0.3 - 2025-03-10
+
+- Automatically load presets from the YTPTube instance.
+- Once the preset is selected, update the default preset in the extension options.
+
 ## 0.0.1 - 2025-03-08
 
 - Initial release version